update according to merge and move editors swr
